\name{graphml-package}
\alias{graphml-package}
\alias{graphml}
\docType{package}
\title{
\packageTitle{graphml}
}
\description{
\packageDescription{graphml}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{graphml}
\packageIndices{graphml}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{graphml}

Maintainer: \packageMaintainer{graphml}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
